+++ library/src/main/java/com/cookieinformation/mobileconsents/TextTranslation.kt
@@ -3,11 +3,9 @@ package com.cookieinformation.mobileconsents
 /**
  * Translation of [ConsentItem], with short and long text of consent (one of them cannot be empty).
  * @param languageCode code of transaction's language, e.g. "EN",
- * @param shortText consent's short explanation.
- * @param longText consent's detailed explanation.
+ * @param text text in a given language.
  */
-public data class ConsentTranslation(
+public data class TextTranslation(
   val languageCode: String,
-  val longText: String,
-  val shortText: String
+  val text: String
 )