+++ library/src/main/java/com/cookieinformation/mobileconsents/networking/response/ConsentTranslationResponse.kt
@@ -5,14 +5,14 @@ import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
-internal data class TranslationResponse(
+internal data class ConsentTranslationResponse(
   @Json(name = "language") val language: String,
   @Json(name = "longText") val longText: String,
   @Json(name = "shortText") val shortText: String
 )
 
-internal fun TranslationResponse.toDomain() = ConsentTranslation(
-  language = language,
+internal fun ConsentTranslationResponse.toDomain() = ConsentTranslation(
+  languageCode = language,
   longText = longText,
   shortText = shortText
 )