+++ library/src/main/java/com/cookieinformation/mobileconsents/networking/response/ItemResponse.kt
@@ -5,13 +5,27 @@ import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import java.util.UUID
 
+private const val typeSetting = "setting"
+private const val typeInfo = "info"
+
 @JsonClass(generateAdapter = true)
 internal data class ItemResponse(
-  @Json(name = "translations") val translations: List<TranslationResponse>,
-  @Json(name = "universalConsentItemId") val consentItemId: UUID
+  @Json(name = "translations") val translations: List<ConsentTranslationResponse>,
+  @Json(name = "universalConsentItemId") val consentItemId: UUID,
+  @Json(name = "required") val required: Boolean,
+  @Json(name = "type") val type: String
 )
 
 internal fun ItemResponse.toDomain() = ConsentItem(
-  translations = translations.map(TranslationResponse::toDomain),
-  consentItemId = consentItemId
+  translations = translations.map(ConsentTranslationResponse::toDomain),
+  consentItemId = consentItemId,
+  required = required,
+  type = type.toDomainItemType()
 )
+
+private fun String.toDomainItemType(): ConsentItem.Type =
+  when (this) {
+    typeSetting -> ConsentItem.Type.Setting
+    typeInfo -> ConsentItem.Type.Info
+    else -> ConsentItem.Type.Setting // TODO Ensure if default value should be "SETTING"
+  }